requirements.txt
streamlit==1.32.0
streamlit-chat==0.1.1
openai==1.12.0
google-generativeai==0.3.2
requests==2.31.0
python-dotenv==1.0.1
langchain==0.1.9
langchain-openai==0.0.7
langchain-google-genai==0.0.9
langchain-community==0.0.24
